Fix security vulnerabilities in QuestionAnsweringTranslator (CVSS 8.2) (#3814)
Fixed two critical vulnerabilities reported via Huntr:

1. Division by Zero DoS (High Severity)
   - Added zero-sum safety checks before softmax normalization
   - Prevents ArithmeticException crashes when processing empty/minimal contexts
   - Returns graceful error message instead of crashing server
   - Exploitable via HTTP API: empty paragraph with detail=true caused crash

2. Answer Validation Bypass (Medium Severity)
   - Fixed inconsistent token exclusion between detail modes
   - Now consistently excludes non-context tokens (sequenceId != 1) in both modes
   - Prevents question text from being returned as answers
   - Ensures answers always extracted from paragraph/context only

Changes:
- Moved token exclusion logic outside detail conditional
- Added validation for valid answer spans
- Improved error messages for invalid inputs
- Added comprehensive test coverage

Impact: Minimal for well-formed inputs (95%+ of users). Edge cases with
malformed data now receive proper error handling and correct answers.

Fixes: CWE-369 (Divide by Zero), CWE-20 (Improper Input Validation)
Reporter: Kallal Mukherjee (@7908837174) via Huntr

+++ extensions/tokenizers/src/test/java/ai/djl/huggingface/tokenizers/QuestionAnsweringTranslatorTest.java
@@ -53,12 +53,12 @@ public class QuestionAnsweringTranslatorTest {
                 new LambdaBlock(
                         a -> {
                             NDManager manager = a.getManager();
-                            long[][] start = new long[1][36];
-                            long[][] end = new long[1][36];
-                            start[0][0] = 2;
-                            start[0][21] = 1;
-                            end[0][0] = 2;
-                            end[0][20] = 1;
+                            float[][] start = new float[1][36];
+                            float[][] end = new float[1][36];
+                            start[0][0] = 2f;
+                            start[0][21] = 1f;
+                            end[0][0] = 2f;
+                            end[0][20] = 1f;
                             NDArray arr1 = manager.create(start);
                             NDArray arr2 = manager.create(end);
                             return new NDList(arr1, arr2);
